Временно запускается
1)control_database - надо переписать, пусть контралирует только пуск
форм и генерация нового класса для создания переменных к форме
2)Создать класс переменных к форме (инициализируется перед каждым
вызовом формы)
3)Выяснить надо ли чистить отработавшие переменные(классы)
4)Дописать работу с COM портом
5)Всё хранится в виде файлов, только загружается перед вызовом формы,
наверно так

+++ Akupunctura/Logik/Files/one_dimension.cs
@@ -8,7 +8,7 @@ namespace Akupunctura.Logik.Files
 [Serializable]
     public class one_dimension
     {
-        public double Current = 0;
-        public double Voltage = 0;
+        public Int32 Current = 0;
+        public Int32 Voltage = 0;
     }
 }
